+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Unreleased
 
-### Bug Fixes
-
 ### Compiler
 
 ### Build tool
+
+- The build tool now generates Hexdocs URLs using the new format of
+  `package.hexdocs.pm` rather than `hexdocs.pm/package`.
+  ([Surya Rose](https://github.com/GearsDatapacks))
 
 ### Language server